Added set ambient routine and fixed index buffer allocation bug which incorrectly used DWORD rather than WORD. Modified demo to use UU vertices.

+++ LaputaEngine/LaputaD3D/include/vertices/LptaD3DDynamicBuffer.h
@@ -31,7 +31,7 @@ public:
 
 private:
     bool LockedBuffers(void **vertexWriteBuffer, unsigned int vertexByteSize,
-        DWORD **indexWriteBuffer, unsigned int indexByteSize, DWORD lockFlag);
+        WORD **indexWriteBuffer, unsigned int indexByteSize, DWORD lockFlag);
 
 private:
     const lpta::VERTEX_TYPE vertexType;
